Add ability to have "singleton" tasks.
This avoids race conditions that could happen with the "remove_existing"
functionality by having a unique index on the task_hash and a
"singleton" field that is only not NULL (and therefore can only be
considered a unique row) when "singleton=True" is passed when the task
is scheduled. Then use bulk_create and ignore_conflicts=True when
creating the task so that we just retain the existing task if another
one (scheduled with "singleton=True" and the same task_hash) is
scheduled. This will prevents duplicates in a reliable way.